refactor tests after removing ast_compiler suite
