+++ tests/bitwise.rs
@@ -1,7 +1,11 @@
 #[path = "ast_compiler_helpers.rs"]
 mod ast_compiler_helpers;
 
+#[path = "wasm_harness.rs"]
+mod wasm_harness;
+
 use ast_compiler_helpers::compile_with_ast_compiler;
+use wasm_harness::run_wasm_main;
 use wasmi::{Engine, Linker, Module, Store, TypedFunc};
 
 #[test]
@@ -58,4 +62,34 @@ fn main() -> i32 {
     let main_result = main_fn.call(&mut store, ()).expect("failed to call main");
     let expected_main = ((12 & 5) << 1) + (12 | 5) + ((-8 << 1) + (-8 >> 1));
     assert_eq!(main_result, expected_main);
+}
+
+#[test]
+fn bitwise_expressions_execute() {
+    let source = r#"
+fn evaluate(a: i32, b: i32, shift: i32) -> i32 {
+    let mask: i32 = (a & b) | ((a | b) >> shift);
+    (mask << 1) + (a >> shift)
+}
+
+fn main() -> i32 {
+    let first: i32 = evaluate(29, 23, 2);
+    let second: i32 = evaluate(-64, 7, 3);
+    first + second
+}
+"#;
+
+    let wasm = compile_with_ast_compiler(source);
+    let engine = wasmi::Engine::default();
+    let result = run_wasm_main(&engine, &wasm);
+
+    let expected = {
+        let eval = |a: i32, b: i32, shift: i32| {
+            let mask = (a & b) | ((a | b) >> shift);
+            (mask << 1) + (a >> shift)
+        };
+        eval(29, 23, 2) + eval(-64, 7, 3)
+    };
+
+    assert_eq!(result, expected);
 }
