+++ tests/entry_point.rs
@@ -1,8 +1,7 @@
 #[path = "ast_compiler_helpers.rs"]
 mod ast_compiler_helpers;
 
-use ast_compiler_helpers::{compile_with_ast_compiler, try_compile_with_ast_compiler};
-use wasmi::{Engine, Linker, Module, Store, TypedFunc};
+use ast_compiler_helpers::try_compile_with_ast_compiler;
 
 #[test]
 fn program_requires_main() {
@@ -17,7 +16,6 @@ fn helper() -> i32 {
 }
 
 #[test]
-#[ignore]
 fn main_cannot_accept_parameters() {
     let source = r#"
 fn main(value: i32) -> i32 {
@@ -25,25 +23,9 @@ fn main(value: i32) -> i32 {
 }
 "#;
 
-    let wasm = compile_with_ast_compiler(source);
-
-    let engine = Engine::default();
-    let mut wasm_reader = wasm.as_slice();
-    let module = Module::new(&engine, &mut wasm_reader).expect("failed to create module");
-    let mut store = Store::new(&engine, ());
-    let linker = Linker::new(&engine);
-    let instance = linker
-        .instantiate(&mut store, &module)
-        .expect("failed to instantiate module")
-        .start(&mut store)
-        .expect("failed to start module");
-
-    let main: TypedFunc<i32, i32> = instance
-        .get_typed_func(&mut store, "main")
-        .expect("expected exported main");
-
-    let result = main.call(&mut store, 42).expect("failed to call main");
-    assert_eq!(result, 42);
+    let error = try_compile_with_ast_compiler(source)
+        .expect_err("expected main with parameters to be rejected");
+    assert!(error.produced_len <= 0);
 }
 
 #[test]
@@ -56,5 +38,26 @@ fn main() -> bool {
 "#;
 
     let error = try_compile_with_ast_compiler(source).expect_err("expected main return type error");
+    assert!(error.produced_len <= 0);
+}
+
+#[test]
+fn main_function_name_must_be_unique() {
+    let source = r#"
+fn main() -> i32 {
+    1
+}
+
+fn helper() -> i32 {
+    2
+}
+
+fn main() -> i32 {
+    3
+}
+"#;
+
+    let error = try_compile_with_ast_compiler(source)
+        .expect_err("programs should not allow multiple main functions");
     assert!(error.produced_len <= 0);
 }
